requirements.txt
numpy
pandas
seaborn
plotly.express